requirements.txt
discord.py
pydub
ffmpeg-python
requests
python-dotenv
notion-client
transformers
torch
sentencepiece
